[5875] added tests and contract creation refactoring
